service-workers/service-worker/navigation-preload/samesite-cookies.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Navigation Preload: SameSite cookies</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/test-helpers.sub.js"></script>
<body>
<script>
const scope = 'resources/set-cookie.py';
const script = 'resources/samesite-cookies-worker.js';

async function samesite_cookies_test(t, samesite) {
  const frame1 = await with_iframe(scope + '?samesite=' + samesite);
  // The server sends the value of "dummy-cookie-key" if it exists, otherwise sends 0.
  assert_equals(frame1.contentDocument.body.textContent, '0');
  t.add_cleanup(() => frame1.remove());

  // Second navigation request to the server handled by navigation preload with cookies.
  const frame2 = await with_iframe(scope + '?samesite=' + samesite);
  // The server sends the value of "dummy-cookie-key" if it exists, otherwise sends 0.
  assert_equals(frame2.contentDocument.body.textContent, '1');
  t.add_cleanup(() => frame2.remove());
}

promise_test(async (t) => {
  const registration =
    await service_worker_unregister_and_register(t, script, scope);
  await wait_for_state(t, registration.installing, 'activated');
  await registration.navigationPreload.enable();

  promise_test(async () => {
    await samesite_cookies_test(t, 'None');
    registration.unregister();
  }, 'Navigation Preload for same site cookies (None).');

  promise_test(async () => {
    await samesite_cookies_test(t, 'Strict');
    registration.unregister();
  }, 'Navigation Preload for same site cookies (Strict).');

  promise_test(async () => {
    await samesite_cookies_test(t, 'Lax');
    registration.unregister();
  }, 'Navigation Preload for same site cookies (Lax).');
}, 'Set up a service worker for navigation preload tests.');

</script>
</body>
